cast in computeDate

+++ src/Ticket.php
@@ -445,7 +445,7 @@ class Ticket extends CommonITILObject
                 $data["slalevels_id_ttr"] = SlaLevel::getFirstSlaLevel($slas_id);
             }
            // Compute time_to_resolve
-            $data['sla_waiting_duration'] = (int) ($this->fields['sla_waiting_duration'] ?: 0);
+            $data['sla_waiting_duration'] = $this->fields['sla_waiting_duration'] ?: 0;
             $data[$dateField]             = $sla->computeDate($date, (int) $data['sla_waiting_duration']);
         } else {
             $data["slalevels_id_ttr"]     = 0;
